+++ frontend/models/Request.java
@@ -1,12 +1,12 @@
 package models;
 
 public class Request {
-    private int userId;
+    private Long userId;
     private String name;
     private String avatarUrl;
     private String sentAt;
 
-    public Request(String name, String avatarUrl, int userId, String sendTime) {
+    public Request(String name, String avatarUrl, Long userId, String sendTime) {
         this.name = name;
         this.avatarUrl = avatarUrl;
         this.userId = userId;
@@ -33,11 +33,11 @@ public class Request {
         return sentAt;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
